fixed serialport constructor for connection
