+++ desktop/usb-con/index.html
@@ -12,8 +12,9 @@
   <body>
     <h1>Connected USB Devices</h1>
     <div id="error"></div>
+    <div id="receivedData"></div>
     <table id="ports"></table>
     
   </body>
-  <script src="renderer.js"></script>
+  <script src="./src/renderer/renderer.js"></script>
 </html>
