+++ desktop/usb-con/package.json
@@ -1,7 +1,7 @@
 {
   "name": "usb-con",
   "version": "0.1.0",
-  "main": "index.js",
+  "main": "./src/main/main.js",
   "homepage": "./",
   "dependencies": {
     "@testing-library/jest-dom": "^5.17.0",
@@ -9,7 +9,7 @@
     "@testing-library/user-event": "^13.5.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-scripts": "^3.0.1",
+    "react-scripts": "^5.0.1",
     "serialport": "^12.0.0",
     "usb": "^2.11.0",
     "web-vitals": "^2.1.4",
