+++ desktop/usb-con/src/main/preload.js
@@ -1,0 +1,12 @@
+const { ipcRenderer } = require('electron');
+
+window.addEventListener('DOMContentLoaded', () => {
+    const deviceList = document.getElementById('deviceList');
+
+
+    ipcRenderer.send('getDevices');
+    
+    ipcRenderer.on('devices', (_event, devices) => {
+        updateDeviceList(devices);
+    });
+});
